# Improve String.Substring performance