WIP just to get the ball rolling regarding discussion of changing exception messages or similar. cc @adamsitnik @GrabYourPitchforks @stephentoub 

Change to `Substring(int startIndex)` improvements (note highly volatile). This comes with all the caveats of potential increases code size etc.

``` ini

BenchmarkDotNet=v0.13.1.1620-nightly, OS=Windows 10.0.19044.1348 (21H2)
AMD Ryzen 9 5950X, 1 CPU, 32 logical and 16 physical cores
.NET SDK=7.0.100-alpha.1.21568.2
  [Host]     : .NET 6.0.0 (6.0.21.48005), X64 RyuJIT
  Job-FWUOEW : .NET 7.0.0 (42.42.42.42424), X64 RyuJIT
  Job-HPGSVV : .NET 7.0.0 (42.42.42.42424), X64 RyuJIT

PowerPlanMode=00000000-0000-0000-0000-000000000000  Arguments=/p:DebugType=portable,-bl:benchmarkdotnet.binlog  IterationTime=250.0000 ms  
MaxIterationCount=20  MinIterationCount=15  WarmupCount=1  

```
|        Method |        Job |                                                                                                    Toolchain |                    s |  i |       Mean |     Error |    StdDev |     Median |       Min |        Max | Ratio | RatioSD |  Gen 0 | Allocated |
|-------------- |----------- |------------------------------------------------------------------------------------------------------------- |--------------------- |--- |-----------:|----------:|----------:|-----------:|----------:|-----------:|------:|--------:|-------:|----------:|
| **Substring_Int** | **Job-FWUOEW** |  **m** | **dzsdzsDDZSDZSDZSddsz** |  **0** |  **1.1022 ns** | **0.0068 ns** | **0.0064 ns** |  **1.1049 ns** | **1.0853 ns** |  **1.1092 ns** |  **1.00** |    **0.00** |      **-** |         **-** |
| Substring_Int | Job-HPGSVV | pr | dzsdzsDDZSDZSDZSddsz |  0 |  0.4543 ns | 0.0069 ns | 0.0065 ns |  0.4541 ns | 0.4382 ns |  0.4658 ns |  0.41 |    0.01 |      - |         - |
|               |            |                                                                                                              |                      |    |            |           |           |            |           |            |       |         |        |           |
| **Substring_Int** | **Job-FWUOEW** |  **m** | **dzsdzsDDZSDZSDZSddsz** |  **7** |  **8.6245 ns** | **2.2287 ns** | **2.5666 ns** |  **9.5946 ns** | **5.4283 ns** | **12.5358 ns** |  **1.00** |    **0.00** | **0.0029** |      **48 B** |
| Substring_Int | Job-HPGSVV | pr | dzsdzsDDZSDZSDZSddsz |  7 |  7.1005 ns | 1.1957 ns | 1.3770 ns |  6.5014 ns | 5.6306 ns |  9.5741 ns |  0.88 |    0.22 | 0.0029 |      48 B |
|               |            |                                                                                                              |                      |    |            |           |           |            |           |            |       |         |        |           |
| **Substring_Int** | **Job-FWUOEW** |  **m** | **dzsdzsDDZSDZSDZSddsz** | **10** | **10.3062 ns** | **0.6670 ns** | **0.7681 ns** | **10.5393 ns** | **8.8757 ns** | **11.3953 ns** |  **1.00** |    **0.00** | **0.0029** |      **48 B** |
| Substring_Int | Job-HPGSVV | pr | dzsdzsDDZSDZSDZSddsz | 10 |  7.5328 ns | 0.6341 ns | 0.7302 ns |  7.4150 ns | 6.3654 ns |  9.1740 ns |  0.73 |    0.07 | 0.0028 |      48 B |
|               |            |                                                                                                              |                      |    |            |           |           |            |           |            |       |         |        |           |
| **Substring_Int** | **Job-FWUOEW** |  **m** | **dzsdzsDDZSDZSDZSddsz** | **19** | **12.9457 ns** | **1.6428 ns** | **1.8919 ns** | **13.3785 ns** | **5.5022 ns** | **14.3101 ns** |  **1.00** |    **0.00** | **0.0014** |      **24 B** |
| Substring_Int | Job-HPGSVV | pr | dzsdzsDDZSDZSDZSddsz | 19 |  6.8188 ns | 0.9429 ns | 1.0858 ns |  6.6514 ns | 5.0843 ns |  9.0324 ns |  0.55 |    0.18 | 0.0014 |      24 B |